requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==3.2
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
mysqlclient==2.1.1
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
